refactor(web-app): move ToastProvider into providers directory
