+++ docker/web-app/src/app/[tenant]/__tests__/settings.test.tsx
@@ -6,7 +6,7 @@ import assert from 'node:assert';
 import test from 'node:test';
 import { renderToString } from 'react-dom/server';
 import SettingsPage from '../settings/page';
-import ToastProvider from '../../../components/providers/ToastProvider';
+import ToastProvider from '../../../providers/ToastProvider';
 
 test('SettingsPage shows general tab', () => {
   const html = renderToString(
