+++ docker/web-app/src/app/__tests__/account.test.tsx
@@ -6,7 +6,7 @@ import assert from 'node:assert';
 import test from 'node:test';
 import { renderToString } from 'react-dom/server';
 import AccountPage from '../account/page';
-import ToastProvider from '../../components/providers/ToastProvider';
+import ToastProvider from '../../providers/ToastProvider';
 
 test('AccountPage renders without crashing', () => {
   const html = renderToString(
