+++ docker/web-app/src/components/settings/GeneralBrandingForm.tsx
@@ -7,7 +7,7 @@ import * as React from 'react';
 import { Button, Stack, TextField, MenuItem } from '@mui/material';
 import { SettingsCard } from '../common/SettingsCard';
 import { api } from '../../lib/api/client';
-import { useToast } from '../providers/ToastProvider';
+import { useToast } from '../../providers/ToastProvider';
 
 export function GeneralBrandingForm({ tenant }: { tenant: string }) {
   const { toast } = useToast();
