+++ docker/web-app/src/providers/AppToast.tsx
@@ -1,6 +1,6 @@
 'use client';
 import * as React from 'react';
-import ToastProvider from '../components/providers/ToastProvider';
+import ToastProvider from './ToastProvider';
 
 // Small wrapper if you prefer to mount globally in src/app/layout.tsx
 export default function AppToast({ children }: { children: React.ReactNode }) {
